moved function from server into helper file
